+++ frontend/frontend/components/index.tsx
@@ -10,7 +10,7 @@ export { default as DragAndDropContainer } from './upload/DragAndDropContainer';
 export { default as SettingsDropdown } from './dropdowns/SettingsDropdown';
 
 // Toggle Components
-export { default as ViewToggle } from './toggle/ViewToggle';
+export { default as ViewToggle } from './menus/ViewToggle';
 
 // Cycles Components
 export { default as CyclesTopUpComponent } from './cycles/CyclesTopUpButton';
@@ -20,3 +20,6 @@ export { default as GetBalancesComponent } from './cycles/GetBalancesButton';
 export { default as LoadingOverlay } from './overlays/LoadingOverlay/LoadingOverlay';
 export { default as ErrorNotification } from './overlays/ErrorNotification';
 export { default as DeleteConfirmation } from './overlays/DeleteConfirmation/DeleteConfirmation';
+
+// Copy Components
+export { default as CopyToClipboard } from './copy/CopyToClipboard';
