+++ frontend/frontend/components/menus/ViewToggle.tsx
@@ -23,16 +23,16 @@ const ViewToggle: React.FC<ViewToggleProps> = ({ viewMode, setViewMode }) => {
         Document Store
       </button>
       <button
-        className={`view-toggle-button ${viewMode === 'json' ? 'active' : ''}`}
-        onClick={() => setViewMode('json')}
-      >
-        Stable Memory
-      </button>
-      <button
         className={`view-toggle-button ${viewMode === 'public' ? 'active' : ''}`}
         onClick={() => setViewMode('public')}
       >
         Public Data
+      </button>
+      <button
+        className={`view-toggle-button ${viewMode === 'json' ? 'active' : ''}`}
+        onClick={() => setViewMode('json')}
+      >
+        Stable Memory
       </button>
       <button
         className={`view-toggle-button ${viewMode === 'admin' ? 'active' : ''}`}
