+++ frontend/frontend/utils/jsonSyntaxHighlight.tsx
@@ -1,0 +1,20 @@
+// utils/jsonSyntaxHighlight.ts
+export const syntaxHighlight = (json: string) => {
+    if (!json) return "";
+    return json
+      .replace(/(&)/g, '&amp;')
+      .replace(/(>)/g, '&gt;')
+      .replace(/(<)/g, '&lt;')
+      .replace(/("(\\u[\da-fA-F]{4}|\\[^u]|[^\\"])*?"(\s*:)?|\b(true|false|null)\b|\d+)/g, (match) => {
+        let cls = "number";
+        if (/^"/.test(match)) {
+          cls = /:$/.test(match) ? "key" : "string";
+        } else if (/true|false/.test(match)) {
+          cls = "boolean";
+        } else if (/null/.test(match)) {
+          cls = "null";
+        }
+        return `<span class="${cls}">${match}</span>`;
+      });
+  };
+
